third_party/WebKit/LayoutTests/platform/win/accessibility/aria-owns-dynamic-changes-expected.txt

This is a testharness.js-based test.
PASS Aria-owns is dynamically recomputed. 
PASS A parent can use aria-owns to reorder its children into a more logical AX ordering. 
Harness: the test ran to completion.
